Updated scalar processing to reflect new input units
The pft weights are now provided as fraction of total grid cell.
The scalar calculations have been updated to handle this, but have
not been tested. Also, the control for applying scalars has been
moved into this code so that scalars can be calculated and
written even if they are not applied.

	modified:   cpl/include/GCAM_E3SM_interface.h
	modified:   cpl/include/carbon_scalers.h
	modified:   cpl/source/GCAM_E3SM_interface.cpp
	modified:   cpl/source/GCAM_E3SM_interface_wrapper.cpp
	modified:   cpl/source/carbon_scalers.cpp